Apparently, the version of the `glm` we used was weird
The version I am using now (downloaded from
https://www.cs.rit.edu/usr/local/pub/ncs/graphics/OpenGL/SIGGRAPHTutorials/glm.c )
seems to be working.

(it is not clear though. I am having these SegFaults and it can even be
that the problem is related to the GLM -- though it is not clear how.
The SegFaults stopped after I replaced the old version of the glm with
this, so I am guessing there is something different -- and better -- with
this new version)

 Changes to be committed:
	modified:   lib/glm.cpp
	modified:   lib/glm.h