Move code to a `src` folder. Create README.md
The README.md still needs a lot of improvement; but at least now the
repository won't look that ugly. The `src` folder contains the most
important code, but the `lib` folder contains other libraries we
had found in the internet in the times when we worked on this code.

I don't like the fact that we don't know where we took all those
.obj from. We should check for their licence accordingly. I think
we should probably remove all of them and use some others that we
are sure are free (i.e., whose licence is anything like CC, or GPL,
or MIT, ...). I'll probably do that during the weekend.

 Changes to be committed:
	modified:   Makefile
	renamed:    Camera.cpp -> src/Camera.cpp
	renamed:    Camera.h -> src/Camera.h
	renamed:    Game.cpp -> src/Game.cpp
	renamed:    Game.h -> src/Game.h
	renamed:    Level.cpp -> src/Level.cpp
	renamed:    Level.h -> src/Level.h
	renamed:    LightInfo.cpp -> src/LightInfo.cpp
	renamed:    LightInfo.h -> src/LightInfo.h
	renamed:    Material.cpp -> src/Material.cpp
	renamed:    Material.h -> src/Material.h
	renamed:    Object.cpp -> src/Object.cpp
	renamed:    Object.h -> src/Object.h
	renamed:    ObjectBall.cpp -> src/ObjectBall.cpp
	renamed:    ObjectBall.h -> src/ObjectBall.h
	renamed:    ObjectModel.cpp -> src/ObjectModel.cpp
	renamed:    ObjectModel.h -> src/ObjectModel.h
	renamed:    ObjectStick.cpp -> src/ObjectStick.cpp
	renamed:    ObjectStick.h -> src/ObjectStick.h
	renamed:    constants.h -> src/constants.h
	renamed:    opengl.cpp -> src/opengl.cpp
	renamed:    utils.cpp -> src/utils.cpp
	renamed:    utils.h -> src/utils.h